+++ packages/ad/placeholder.js
@@ -1,11 +1,8 @@
 import React from "react";
-import { StyleSheet, View, Text, Platform, ViewPropTypes } from "react-native";
+import { StyleSheet, View, Text, ViewPropTypes } from "react-native";
 import PropTypes from "prop-types";
 
 import TimesWatermark from "./ad-watermark";
-
-const fontFamilyWebAndIos = "TimesDigitalW04";
-const fontFamilyAndroid = "TimesDigitalW04-Regular";
 
 const getStyles = (width, height) =>
   StyleSheet.create({
@@ -26,8 +23,7 @@ const getStyles = (width, height) =>
     },
     placeholderText: {
       fontSize: 11,
-      fontFamily:
-        Platform.OS === "android" ? fontFamilyAndroid : fontFamilyWebAndIos,
+      fontFamily: "TimesDigitalW04",
       backgroundColor: "#F9F9F9",
       color: "#696969",
       borderColor: "#DBDBDB",