+++ packages/article-summary/article-summary.js
@@ -1,12 +1,9 @@
 import React from "react";
-import { Text, View, Platform } from "react-native";
+import { Text, View } from "react-native";
 import PropTypes from "prop-types";
 import { renderTrees, treePropType } from "@times-components/markup";
 import DatePublication from "@times-components/date-publication";
 import renderer from "./article-summary-renderer";
-
-const fontFamilyWebAndIos = "TimesDigitalW04";
-const fontFamilyAndroid = "TimesDigitalW04-Regular";
 
 const styles = {
   container: {},
@@ -28,8 +25,7 @@ const styles = {
   text: {
     color: "#696969",
     fontSize: 14,
-    fontFamily:
-      Platform.OS === "android" ? fontFamilyAndroid : fontFamilyWebAndIos,
+    fontFamily: "TimesDigitalW04",
     lineHeight: 20,
     marginBottom: 10,
     flexWrap: "wrap"