+++ packages/article/styles/article-style/native.js
@@ -2,7 +2,6 @@ import sharedStyles from "./shared";
 
 const nativeStyles = {
   articleTextElement: Object.assign({}, sharedStyles.articleTextElement, {
-    fontStyle: "normal",
     marginBottom: 20
   })
 };