+++ packages/article/styles/const/index.android.js
@@ -1,7 +1,0 @@
-import sharedConst from "./shared";
-
-const androidConst = Object.assign({}, sharedConst, {
-  fontFamilyBody: "TimesDigitalW04-Regular"
-});
-
-export default androidConst;