+++ packages/article/styles/const/index.js
@@ -1,3 +1,7 @@
-import sharedConst from "./shared";
-
-export default sharedConst;
+export default {
+  fontFamilyBody: "TimesDigitalW04",
+  fontFamilyStandFirst: "TimesModern-Regular",
+  colourText: "#333333",
+  colourHeadline: "#000000",
+  colourBorder: "#D0CECE"
+};