+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.23.2"></a>
+## [0.23.2](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.23.1...@times-components/author-profile@0.23.2) (2017-11-07)
+
+
+
+
+<a name="0.23.1"></a>
+## [0.23.1](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.23.0...@times-components/author-profile@0.23.1) (2017-11-07)
+
+
+
+
 <a name="0.23.0"></a>
 # [0.23.0](https://github.com/newsuk/times-components/compare/@times-components/author-profile@0.22.5...@times-components/author-profile@0.23.0) (2017-11-07)