+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.2"></a>
+## 0.2.2 (2017-11-07)
+
+
+### Bug Fixes
+
+* article headline title ([#330](https://github.com/newsuk/times-components/issues/330)) ([16c71fe](https://github.com/newsuk/times-components/commit/16c71fe))
+
+
+
+
 <a name="0.2.1"></a>
 ## [0.2.1](https://github.com/newsuk/times-components/compare/@times-components/tracking@0.2.0...@times-components/tracking@0.2.1) (2017-10-24)